requirements.txt
transformers
torch
SentencePiece
python-dotenv
sounddevice
google-generativeai
protobuf==4.25.3